backend/requirements.txt
setuptools>=68.0.0
wheel>=0.41.0
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6
pydantic>=2.3.0
chromadb>=0.4.18
numpy>=1.25.2
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
aiofiles>=23.2.1
jinja2>=3.1.2
typing-extensions>=4.7.0
uuid>=1.30
watchdog>=3.0.0

# Processamento de texto e embeddings
transformers>=4.34.1
torch>=2.0.1
torchvision>=0.15.2
torchaudio>=2.0.2

# Processamento de PDFs
PyMuPDF>=1.22.5  # Para extração de texto de PDFs

# Processamento de imagens
pytesseract>=0.3.10  # OCR para extração de texto de imagens
Pillow>=10.0.1  # Processamento básico de imagens
opencv-python>=4.8.1.78  # Processamento avançado de imagens

# Processamento de áudio e vídeo
openai-whisper>=20231117  # Transcrição de áudio para texto
ffmpeg-python>=0.2.0  # Manipulação de vídeos e extração de áudio
moviepy>=1.0.3  # Processamento de vídeos

# Outras utilidades
scikit-learn>=1.3.0  # Para algoritmos de ML
matplotlib>=3.7.2  # Para visualizações
pandas>=2.0.3  # Para manipulação de dados 